[fix/TK-121]: 여행 홈, 결제내역 경로 수정
