+++ apps/mobile/app/travel/[id]/payments/page.module.css
@@ -1,0 +1,39 @@
+.container {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  align-items: center;
+}
+
+.subContainer {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  margin-top: 16px;
+  /* padding: 0 25px; */
+}
+
+.subContainer > *:first-child {
+  margin-right: auto;
+}
+
+.subContainer > *:nth-child(2) {
+  margin-right: 10px;
+}
+
+.paymentContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 100%;
+  margin: 21px 0 25px 0;
+}
+
+.paymentWrapper {
+  width: 100%;
+}
+
+.date {
+  color: var(--color-gray-500);
+  font-size: 14px;
+}
