--- apps/mobile/app/travel/page.module.css
@@ -1,76 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-}
-
-.subtitleContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  /* flex-wrap: wrap; */
-  gap: 3px;
-}
-
-.subContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 19px;
-  margin-bottom: 28px;
-  padding-top: 35px;
-}
-
-.imgWrapper {
-  width: calc(100% - 100px);
-  /* min-width: 280px; */
-  height: 207px;
-  border-radius: 20px;
-  background: var(--color-gray-100) url('/friends/wibee_gray.png') no-repeat center / 162px 156.67px;
-}
-
-.btnWrapper {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  /* flex-wrap: wrap; */
-  gap: 8px;
-}
-
-.subtitleWrapper {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.tagWrapper {
-  display: flex;
-  align-items: center;
-  gap: 7px;
-}
-
-.date {
-  font-size: 14px;
-  color: var(--color-gray-600);
-}
-
-.subtitle {
-  font-size: 20px;
-  color: var(--color-gray-800);
-  font-weight: 500;
-}
-
-.tag {
-  padding: 7px 10px;
-  background: var(--color-gray-100);
-  border-radius: 7px;
-}
-
-.friendsWrapper {
-  display: flex;
-  align-items: center;
-}
